create email file and job file

+++ database/migrations/2024_05_15_081453_create_team_user_table.php
@@ -12,8 +12,8 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('team_user', function (Blueprint $table) {
-            $table->integer('team_id');
-            $table->integer('user_id');
+            $table->unsignedBigInteger('team_id');
+            $table->unsignedBigInteger('user_id');
             $table->unique(['team_id' , 'user_id']);
         });
     }
